+++ examples/moving_example.rs
@@ -99,7 +99,8 @@ fn spawn_cloth(
                             material: materials.add(StandardMaterial {
                                 base_color_texture: Some(flag_texture),
                                 cull_mode: None, // Option required to render back faces correctly
-                                double_sided: true, // Option required to render back faces correctly
+                                double_sided: true, /* Option required to render back faces
+                                                  * correctly */
                                 ..Default::default()
                             }),
                             transform: Transform::from_xyz(-1.0, 1.0, 1.01),