+++ src/components/collider.rs
@@ -1,9 +1,9 @@
-use bevy::ecs::component::Component;
-use bevy::reflect::Reflect;
+use bevy::{ecs::component::Component, reflect::Reflect};
 
 /// Enables collisions on a cloth entity
 ///
-/// The collisions will be detected through a cuboid shape using the cloth AABB bounding box.
+/// The collisions will be detected through a cuboid shape using the cloth AABB
+/// bounding box.
 #[derive(Debug, Clone, Component, Reflect)]
 pub struct ClothCollider {
     /// offset to apply on collision projected point to prevent clipping
@@ -13,8 +13,8 @@ pub struct ClothCollider {
     /// - 1 meaning velocity is fully applied
     /// - 2 meaning the double velocity is applied
     pub velocity_coefficient: f32,
-    /// Defines the velocity reduction coefficient for dynamic rigibodies colliding with the cloth,
-    /// improving the cloth effect.
+    /// Defines the velocity reduction coefficient for dynamic rigibodies
+    /// colliding with the cloth, improving the cloth effect.
     pub dampen_others: Option<f32>,
 }